general clean up and renaming
