+++ src/components/Header.js
@@ -5,7 +5,7 @@ import { Link as StyledLink, Flex, Box, Image } from '@chakra-ui/core';
 import logo from '../img/logo.svg';
 
 const Header = ({userState}) => {
-  console.log('userState', userState);
+  // console.log('userState', userState);
   return (
     <Flex
       as="header"
