+++ src/components/mobile/Mobile.js
@@ -1,6 +1,6 @@
-import Nav from './Nav';
+import Nav from './NavigationComponents/Nav';
 import React, {useState} from 'react'
-import BotNav from './NavigationComponents/BotNav'
+// import BotNav from './NavigationComponents/BotNav'
 import MobileEvents from './eventComponents/MobileEvents'
 import styled from 'styled-components'
 import Dashboard from './Dashboard'
@@ -9,7 +9,7 @@ const Mobile = () => {
 
     const [NavState, setNavState] = useState(0) // 0 = calendar, 1 = events, 2 = groups
 
-    const[formOpen, setFormOpen] = useState(false);
+    const [formOpen, setFormOpen] = useState(false);
     const [templateFormOpen, setTemplateFormOpen] = useState(false);
     const [conStart, setConStart] = useState("");
     const [conEnd, setConEnd] = useState("");
@@ -22,6 +22,7 @@ const Mobile = () => {
     justify-content: center;
     position: fixed;
     top: 83vh;`
+
     if(NavState === 0){
     return(
 
@@ -29,23 +30,23 @@ const Mobile = () => {
         <div >
             <Dashboard formOpen={formOpen} setFormOpen={setFormOpen} setTemplateFormOpen={setTemplateFormOpen} templateFormOpen={templateFormOpen} conStart={conStart} conEnd={conEnd} summ={summ} selected={selected} setSelected={setSelected} toggleNav={toggleNav} setToggleNav={setToggleNav}/>
             <NavBar>
-            {toggleNav && <BotNav NavState={NavState} setNavState={setNavState}></BotNav>}
+            {toggleNav && <Nav NavState={NavState} setNavState={setNavState} />}
             </NavBar>
         </div>
     )
     } else if(NavState === 1){
         return(
-        <div>hello im events
+        <div>
 
             <MobileEvents setNavState={setNavState} formOpen={formOpen} setFormOpen={setFormOpen}setTemplateFormOpen={setTemplateFormOpen} templateFormOpen={templateFormOpen} setToggleNav={setToggleNav} toggleNav={toggleNav} conStart={conStart} setConStart={setConStart} conEnd={conEnd} setConEnd={setConEnd} summ={summ} setSumm={setSumm} selected={selected} setSelected={setSelected}></MobileEvents>
             <NavBar>
-            <BotNav NavState={NavState} setNavState={setNavState}></BotNav>
+            <Nav NavState={NavState} setNavState={setNavState} />
             </NavBar>
         </div>
         )
     } else if(NavState === 2){
         return(
-            <div>hello im groups
+            <div>
             <Nav NavState={NavState} setNavState={setNavState} />
         </div>
         )
