+++ src/components/mobile/NewInfCal.js
@@ -1,7 +1,7 @@
-import React, { useState, createRef, Fragment, PureComponent } from "react";
+import React, { Fragment } from "react";
 import { FixedSizeList as List } from "react-window";
 import InfiniteLoader from "react-window-infinite-loader";
-import Calendar from './Calendar';
+import Calendar from './MobileCalendar';
 
 
 const LOADING = 1;
