+++ src/components/mobile/eventComponents/NewEventForm.js
@@ -23,7 +23,7 @@ const addTemplate = async (data, { googleId }) => {
         `${process.env.REACT_APP_ENDPOINT_URL}/api/template`,
         template
       );
-      console.log('response.data', response.data)
+    //   console.log('response.data', response.data)
       return response.data;
     } catch (error) {
       console.log(error);
